+++ src/library/plone/BaseFile.ts
@@ -1,9 +1,12 @@
 'use strict';
 import * as vscode from 'vscode';
 import { PloneObject } from ".";
-import { post, getBuffer, get } from '../util';
+import { post, getBuffer } from '../util';
 import { RequestOptions } from 'https';
 
+/**
+ * A Plone Object with data
+ */
 export default abstract class BaseFile extends PloneObject {
 	data: Uint8Array;
 	static readonly fieldname: string;
@@ -35,30 +38,8 @@ export default abstract class BaseFile extends PloneObject {
 		return this.exists = buffer.equals(BaseFile.savedText);
 	}
 
-	load(cookie: string): Promise<boolean> {
-		if (this.loading) {
-			return this.loadingPromise;
-		}
-		this.loading = true;
-		return this.loadingPromise = this._load(cookie);
-	}
-
 	protected async _load(cookie: string): Promise<boolean> {
-		const externalEditPath = this.path.dir + '/externalEdit_/' + this.name;
-		const response = await get({
-			host: this.uri.authority,
-			path: externalEditPath,
-			headers: { cookie },
-		});
-		if (response.statusCode === 302) {
-			this.loading = false;
-			throw vscode.FileSystemError.NoPermissions(this.uri);
-		}
-		else if (response.statusCode !== 200) {
-			this.loading = false;
-			throw vscode.FileSystemError.Unavailable(`${response.statusCode}: ${response.statusMessage}`);
-		}
-		const buffer = await getBuffer(response);
+		const buffer = await this._loadExternalBuffer(cookie);
 		this.data = this.parseExternalEdit(buffer);
 		return this.loaded = true;
 	}