+++ src/library/plone/NewsItem.ts
@@ -1,4 +1,5 @@
-import { Document } from '.';
+import { Page } from '.';
 
-export default class NewsItem extends Document {
+export default class NewsItem extends Page {
+	static readonly type_name = 'News Item';
 }