+++ src/library/plone/Page.ts
@@ -1,0 +1,7 @@
+import { Document } from '.';
+/**
+ * Documents can be checked out
+ */
+export default class Page extends Document {
+	static readonly type_name: string = 'Document';
+}