+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 ## 0.3.0
 
-- Added to More Plone Options context menu: Check out, Cancel Check Out, Check In
+- Added to Plone Options context menu: Check out, Cancel Check Out, Check In
 - Added support for News Items, Events, and Collections (aka Topics)
 - Added setting State
-- Fixed not being able to edit a file's Title/Description
+- Added Title and State to status bar
+- Fixed not being able to edit a file's Title
 
 ## 0.2.1